refactor: #17 - Form 타입 수정
- ID는 Int만 받기 때문에 그렇게 수정

+++ PokeDex_Project/Shared/Model/Varieties/Child/Form.swift
@@ -9,7 +9,7 @@ import Foundation
 
 /// 폼 정보
 struct Form: Codable, Hashable, Equatable {
-    var id:[Double]             // ID
+    var id:[Int]                // ID
     var images: [String]        // 폼 이미지
     var name: [String]          // 폼 이름
 }
